Add conditional support for Network.framework
